x86_64: enable HW_RANDOM and HW_RANDOM_VIRTIO
Enable HW_RANDOM and HW_RANDOM_VIRTIO so the guest can gather entropy
from the host through the virtio-rng device.

This is only enabled on bare x86_64 and not on x86_64-sev, as in the
context of SEV we don't consider the host to be trusted to provide us
with good random numbers.

Signed-off-by: Sergio Lopez <slp@redhat.com>