Makefile: Use aarch64 even on Apple Silicon
On Apple Silicon, `uname -m` returns `arm64` and not `aarch64` (although it does
use `aarch64` when running inside `krunvm`, e.g. with the `build_on_krunvm.sh`
script. Replace `arm64` with `aarch64` in the `ARCH` variable so that the rest
of the makefile has a chance to work.

Note that by default, in the current stage, the `Makefile` does not correctly
build on macOS anyway, since building the Linux kernel on macOS is a bit difficult.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>