Rebase on a 6.1.32 kernel
This is a downgrade from the 6.2.x series to the 6.1.x one. The 6.2.x
series is EOL and moving TO 6.3.x requires a significant amount of work
due to critical changes upstream in vsock. Given that the short/mid-term
plan is to drop TSI in favor of paast, it's probably not worth it
investing the effort into updating the TSI series.

Since the 6.1.x series is LTS, we can stay there and receive bug fixes
until the paast is adopted and we can move forward with newer kernels.

Signed-off-by: Sergio Lopez <slp@redhat.com>