Rebase on a 6.12.3 kernel
Rebase on a 6.12.3 kernel, jumping to the next LTS series. Only a
surprisingly small amount of changes needed in the patches, mostly in
the DGRAM and TSI sets. A new qboot binary was also needed with a very
small change to account for the new SEV kernel entry point (tested under
SEV-ES on Naples).

We also dropped "0009-tests-add-vsock-dgram-tests.patch" to reduce the
maintenance burden (was useless anyway).

Signed-off-by: Sergio Lopez <slp@redhat.com>