utils: Add kernel_size_time.sh to measure size and boot time
This is what I've been using to measure size and execution times for
a test command (based on muvm) for the previous patches. It's made to
be used in a rebase script, for example:

  x ./utils/kernel_size_time.sh
  pick 12181bf config-libkrunfw_x86_64: Disable processor quirks and features we don't need
  x ./utils/kernel_size_time.sh
  pick 5c5badc config-libkrunfw_x86_64: Drop power management features and cpufreq
  x ./utils/kernel_size_time.sh

it's a bit bigger than what I wanted because of the time -p trick:
nowadays most distributions don't ship /usr/bin/time (at least by
default), but many shells skip support for the POSIX compatibility
mode (-p, which simplifies calculations) in their 'time' built-in,
so, if we need it, we need to re-execute under Bash (assuming it's
not the default non-interactive shell).

Signed-off-by: Stefano Brivio <sbrivio@redhat.com>